Refactor UserController
